🎨 [FE] chore: content component 분리
